+++ benchopt/plotting/generate_html.py
@@ -1,20 +1,17 @@
 import json
 import shutil
-import itertools
 import webbrowser
 from pathlib import Path
 from datetime import datetime
-
+import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
 from mako.template import Template
 
 from ..constants import PLOT_KINDS
-from .plot_bar_chart import plot_bar_chart  # noqa: F401
-from .plot_objective_curve import plot_objective_curve  # noqa: F401
-from .plot_objective_curve import plot_suboptimality_curve  # noqa: F401
-from .plot_objective_curve import plot_relative_suboptimality_curve  # noqa: F401 E501
-
+from .plot_bar_chart import computeBarChartData  # noqa: F401
+from .plot_objective_curve import compute_quantiles   # noqa: F401
+from .plot_objective_curve import get_solver_color   # noqa: F401
+from .plot_objective_curve import get_solver_marker  # noqa: F401
 
 ROOT = Path(__file__).parent / "html"
 DEFAULT_HTML_DIR = Path("html")
@@ -24,7 +21,6 @@ FIGURES = "figures"
 TEMPLATE_INDEX = ROOT / "templates" / "index.mako.html"
 TEMPLATE_BENCHMARK = ROOT / "templates" / "benchmark.mako.html"
 TEMPLATE_RESULT = ROOT / "templates" / "result.mako.html"
-TEMPLATE_LOCAL_RESULT = ROOT / "templates" / "local_result.mako.html"
 
 SYS_INFO = {
     "main": [('system-cpus', 'cpu'),
@@ -41,119 +37,13 @@ SYS_INFO = {
 }
 
 
-def generate_plot_benchmark(df, kinds, fname, fig_dir, benchmark_name):
-    """Generate all possible plots for a given benchmark run.
-
-    Parameters
-    ----------
-    df : instance of pandas.DataFrame
-        The benchmark results.
-    kinds : list of str
-        List of the kind of plots that will be generated. This needs to be a
-        sub-list of PLOT_KINDS.keys().
-    fname: str
-        CSV file name.
-    fig_dir : Path
-        Base directory to save figures.
-    benchmark_name : str
-        Name of the benchmark, to prefix the file names.
-
-    Returns
-    -------
-    dict
-        The figures for convergence curves and bar charts
-        for each dataset.
-    """
-    dataset_names = df['data_name'].unique()
-    objective_names = df['objective_name'].unique()
-    obj_cols = [
-        k for k in df.columns
-        if k.startswith('objective_') and k != 'objective_name'
-    ]
-
-    figures = {}
-    n_figure = 0
-    for data_name in dataset_names:
-        figures[data_name] = {}
-        df_data = df[df['data_name'] == data_name]
-        for objective_name in objective_names:
-            figures[data_name][objective_name] = {c: {} for c in obj_cols}
-            df_obj = df_data[df_data['objective_name'] == objective_name]
-
-            for k, obj_col in itertools.product(kinds, obj_cols):
-                if k not in PLOT_KINDS:
-                    raise ValueError(
-                        f"Requesting invalid plot '{k}'. Should be in:\n"
-                        f"{PLOT_KINDS}")
-                plot_func = globals()[PLOT_KINDS[k]]
-                try:
-                    fig = plot_func(df_obj, obj_col=obj_col, plotly=True)
-                    if PLOT_KINDS[k] == "plot_bar_chart":
-                        fig.update_layout(autosize=False,
-                                          width=900,
-                                          height=650)
-                    else:
-                        if len(df_obj["solver_name"].unique()) < 10:
-                            fact_ = 10
-                        else:
-                            fact_ = 100
-                        height = 700 + fact_ * len(objective_names)
-                        fig.update_layout(legend={"xanchor": "center",
-                                                  "yanchor": "top",
-                                                  "y": -.2,
-                                                  "x": .5
-                                                  },
-                                          autosize=False,
-                                          width=900,
-                                          height=height
-                                          )
-                except TypeError:
-                    fig = plot_func(df_obj, obj_col=obj_col)
-                figures[data_name][objective_name][obj_col][k] = export_figure(
-                    fig, f"{benchmark_name}_{fname.name}_{n_figure}", fig_dir
-                )
-                n_figure += 1
-
-    return dict(
-        figures=figures, dataset_names=dataset_names, fname_short=fname.name,
-        objective_names=objective_names, obj_cols=obj_cols, kinds=list(kinds)
-    )
-
-
-def export_figure(fig, fig_name, fig_dir):
-    """Export a figure to HTML or svg.
-
-    Parameters
-    ----------
-    fig : plotly.graph_objs.Figure
-        Figure from plotly
-    fig_name : str
-        Name to be given to the figure.
-    fig_dir : Path
-        Base directory to save figures.
-
-    Returns
-    -------
-    save_name : str
-        Path to the saved figure.
-    """
-    if hasattr(fig, 'to_html'):
-        return fig.to_html(include_plotlyjs=False)
-
-    fig_basename = f"{fig_name}.svg"
-    save_name = fig_dir / fig_basename
-    fig.savefig(save_name)
-    plt.close(fig)
-    return str(save_name)
-
-
 def get_results(fnames, kinds, root_html, benchmark_name, copy=False):
-    """Generate figures from a list of csv files.
+    """Generate figures from a list of result files.
 
     Parameters
     ----------
     fnames : list of Path
-        list of csv files containing the benchmark results.
+        list of result files containing the benchmark results.
     kinds : list of str
         List of the kind of plots that will be generated. This needs to be a
         sub-list of PLOT_KINDS.keys().
@@ -172,38 +62,137 @@ def get_results(fnames, kinds, root_html, benchmark_name, copy=False):
         generated figures.
     """
     results = []
-    fig_dir = root_html / FIGURES
     out_dir = root_html / OUTPUTS
 
     for fname in fnames:
         print(f"Processing {fname}")
 
-        df = pd.read_csv(fname)
+        if fname.suffix == '.parquet':
+            df = pd.read_parquet(fname)
+        else:
+            df = pd.read_csv(fname)
+
         datasets = list(df['data_name'].unique())
         sysinfo = get_sysinfo(df)
-        # Copy CSV if necessary and give a relative path for HTML page access
+        # Copy result file if necessary
+        # and give a relative path for HTML page access
         if copy:
             fname_in_output = out_dir / f"{benchmark_name}_{fname.name}"
             shutil.copy(fname, fname_in_output)
             fname = fname_in_output
-        fname = fname.relative_to(root_html)
+        fname = fname.absolute().relative_to(root_html.absolute())
 
         # Generate figures
         result = dict(
-            fname=fname, datasets=datasets, sysinfo=sysinfo,
-            **generate_plot_benchmark(
-                df, kinds, fname, fig_dir, benchmark_name
-            )
+            fname=fname,
+            fname_short=fname.name,
+            datasets=datasets,
+            sysinfo=sysinfo,
+            dataset_names=df['data_name'].unique(),
+            objective_names=df['objective_name'].unique(),
+            obj_cols=[k for k in df.columns if k.startswith('objective_')
+                      and k != 'objective_name'],
+            kinds=list(kinds),
         )
         results.append(result)
 
     for result in results:
+        html_file_name = f"{result['fname_short'].replace('.csv', '.html')}"
+        html_file_name = f"{html_file_name.replace('.parquet', '.html')}"
+
         result['page'] = (
             f"{benchmark_name}_"
-            f"{result['fname_short'].replace('.csv', '.html')}"
+            f"{html_file_name}"
         )
 
+        # JSON
+        result['json'] = json.dumps(shape_datasets_for_html(df))
+
+        # Assets
+        assets = ['result.js', 'symbols.js', 'main.css', 'hover_index.css',
+                  'arrow_left.svg', 'home.svg', 'download.svg']
+        for asset in assets:
+            with open(root_html / 'static' / asset) as asset_file:
+                result[asset] = asset_file.read()
+
     return results
+
+
+def shape_datasets_for_html(df):
+    """Return a dictionary with plotting data for each dataset."""
+    datasets_data = {}
+
+    for dataset in df['data_name'].unique():
+        datasets_data[dataset] = shape_objectives_for_html(df, dataset)
+
+    return datasets_data
+
+
+def shape_objectives_for_html(df, dataset):
+    """Return a dictionary with plotting data for each objective."""
+    objectives_data = {}
+
+    for objective in df['objective_name'].unique():
+        objectives_data[objective] = shape_objectives_columns_for_html(
+            df, dataset, objective)
+
+    return objectives_data
+
+
+def shape_objectives_columns_for_html(df, dataset, objective):
+    """Return a dictionary with plotting data for each objective column."""
+    objective_columns_data = {}
+    columns = [
+        c for c in df.columns
+        if c.startswith('objective_') and c != 'objective_name'
+    ]
+
+    for column in columns:
+        df_filtered = df.query(
+            "data_name == @dataset & objective_name == @objective"
+        )
+        objective_columns_data[column] = {
+            'solvers': shape_solvers_for_html(df_filtered, column),
+            # Values used in javascript to do computation
+            'transformers': {
+                'c_star': float(df_filtered[column].min() - 1e-10),
+                'max_f_0': float(
+                    df_filtered[df_filtered['stop_val'] == 1][column].max()
+                )
+            }
+        }
+
+    return objective_columns_data
+
+
+def shape_solvers_for_html(df, objective_column):
+    """Return a dictionary with plotting data for each solver."""
+    solver_data = {}
+    for solver in df['solver_name'].unique():
+        df_filtered = df.query("solver_name == @solver")
+
+        # remove infinite values
+        df_filtered = df_filtered.replace([np.inf, -np.inf], np.nan)
+        df_filtered = df_filtered.dropna(subset=[objective_column])
+
+        q1, q9 = compute_quantiles(df_filtered)
+        solver_data[solver] = {
+            'scatter': {
+                'x': df_filtered.groupby('stop_val')['time']
+                                .median().tolist(),
+                'y': df_filtered.groupby('stop_val')[objective_column]
+                                .median().tolist(),
+                'q1': q1.tolist(),
+                'q9': q9.tolist(),
+            },
+            'bar': {
+                **computeBarChartData(df, objective_column, solver)
+            },
+            'color': get_solver_color(solver),
+            'marker': get_solver_marker(solver)
+        }
+
+    return solver_data
 
 
 def get_sysinfo(df):
@@ -238,6 +227,7 @@ def get_sysinfo(df):
                     str(df["platform-architecture"].unique()[0])
                 )
             else:
+                df['version-numpy'] = df['version-numpy'].astype(str)
                 val = df[key].unique()[0]
                 if not pd.isnull(val):
                     return str(val)
@@ -439,6 +429,7 @@ def plot_benchmark_html(fnames, benchmark, kinds, display=True):
     with open(bench_index, "w", encoding="utf-8") as f:
         f.write(rendered)
 
+    print("Rendering benchmark results...")
     # Display the file in the default browser
     if display:
         result_filename = (root_html / results[-1]['page']).absolute()
@@ -447,7 +438,7 @@ def plot_benchmark_html(fnames, benchmark, kinds, display=True):
 
 def plot_benchmark_html_all(patterns=(), benchmarks=(), root=None,
                             display=True):
-    """Generate a HTML rerport for multiple benchmarks.
+    """Generate a HTML report for multiple benchmarks.
 
     This utility is the one used to create https://benchopt.github.io/results.
     It will open all benchmarks in `root` and create a website to browse them.
@@ -494,20 +485,25 @@ def plot_benchmark_html_all(patterns=(), benchmarks=(), root=None,
 
         fnames = []
         for p in patterns:
-            fnames += (benchmark / 'outputs').glob(f"{p}.csv")
+            fnames += list(
+                (benchmark / 'outputs').glob(f"{p}.parquet")
+            ) + list((benchmark / 'outputs').glob(f"{p}.csv"))
         fnames = sorted(set(fnames))
         results = get_results(
             fnames, PLOT_KINDS.keys(), root_html, benchmark.name, copy=True
         )
         len_fnames.append(len(fnames))
-        rendered = render_benchmark(
-            results, benchmark.name, static_dir=static_dir
-        )
+        if len(results) > 0:
+            rendered = render_benchmark(
+                results, benchmark.name, static_dir=static_dir
+            )
 
-        benchmark_filename = (root_html / benchmark.name).with_suffix('.html')
-        print(f"Writing {benchmark.name} results to {benchmark_filename}")
-        with open(benchmark_filename, "w") as f:
-            f.write(rendered)
+            benchmark_filename = (
+                root_html / benchmark.name
+            ).with_suffix('.html')
+            print(f"Writing {benchmark.name} results to {benchmark_filename}")
+            with open(benchmark_filename, "w") as f:
+                f.write(rendered)
 
         htmls = render_all_results(
             results, benchmark.name, static_dir=static_dir