+++ benchopt/runner.py
@@ -7,19 +7,15 @@ from joblib import Parallel, delayed
 from .callback import _Callback
 from .benchmark import _check_name_lists
 from .utils.sys_info import get_sys_info
-from .utils.pdb_helpers import exception_handler
 from .utils.files import uniquify_results
-
+from .utils.pdb_helpers import exception_handler
 from .utils.terminal_output import TerminalOutput
-
-# For compat with the lasso benchmark, expose INFINITY in this module.
-# Should be removed once benchopt/benchmark_lasso#55 is merged
-from .stopping_criterion import INFINITY  # noqa: F401
-
 
 ##################################
 # Time one run of a solver
 ##################################
+
+
 def run_one_resolution(objective, solver, meta, stop_val):
     """Run one resolution of the solver.
 
@@ -238,8 +234,8 @@ def run_one_solver(benchmark, dataset, objective, solver, n_repetitions,
 
 
 def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
-                  dataset_names=None, objective_filters=None,
-                  max_runs=10, n_repetitions=1, timeout=100, n_jobs=1,
+                  dataset_names=None, objective_filters=None, max_runs=10,
+                  n_repetitions=1, timeout=100, n_jobs=1, slurm=None,
                   plot_result=True, html=True, show_progress=True, pdb=False,
                   output="None"):
     """Run full benchmark.
@@ -269,6 +265,9 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
         The maximum duration in seconds of the solver run.
     n_jobs : int
         Maximal number of workers to use to run the benchmark in parallel.
+    slurm : Path | None
+        If not None, launch the job on a slurm cluster using the file to get
+        the cluster config parameters.
     plot_result : bool
         If set to True (default), display the result plot and save them in
         the benchmark directory.
@@ -280,8 +279,8 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
     pdb : bool
         It pdb is set to True, open a debugger on error.
     output_name : str
-        Filename for the csv output. If given, the results will
-        be stored at <BENCHMARK>/outputs/<filename>.csv.
+        Filename for the parquet output. If given, the results will
+        be stored at <BENCHMARK>/outputs/<filename>.parquet.
 
     Returns
     -------
@@ -291,7 +290,6 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
         by the objective is not the same for all parameters, the missing data
         is set to `NaN`.
     """
-    print("Benchopt is running")
     output_name = output
 
     # List all datasets, objective and solvers to run based on the filters
@@ -304,14 +302,19 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
         solver_names, forced_solvers, dataset_names, objective_filters,
         output=output
     )
-
-    results = Parallel(n_jobs=n_jobs)(
-        delayed(run_one_solver)(
-            benchmark=benchmark, dataset=dataset, objective=objective,
-            solver=solver, n_repetitions=n_repetitions, max_runs=max_runs,
-            timeout=timeout, force=force, output=output, pdb=pdb
-        ) for dataset, objective, solver, force, output in all_runs
+    common_kwargs = dict(
+        benchmark=benchmark, n_repetitions=n_repetitions, max_runs=max_runs,
+        timeout=timeout, pdb=pdb
     )
+
+    if slurm is not None:
+        from .utils.slurm_executor import run_on_slurm
+        results = run_on_slurm(slurm, run_one_solver, common_kwargs, all_runs)
+    else:
+        results = Parallel(n_jobs=n_jobs)(
+            delayed(run_one_solver)(**common_kwargs, **kwargs)
+            for kwargs in all_runs
+        )
 
     run_statistics = []
     for curve in results:
@@ -323,15 +326,15 @@ def run_benchmark(benchmark, solver_names=None, forced_solvers=None,
         output.savefile_status()
         raise SystemExit(1)
 
-    # Save output in CSV file in the benchmark folder
+    # Save output in parquet file in the benchmark folder
     timestamp = datetime.now().strftime('%Y-%m-%d_%Hh%Mm%S')
     output_dir = benchmark.get_output_folder()
     if output_name == "None":
-        save_file = output_dir / f'benchopt_run_{timestamp}.csv'
+        save_file = output_dir / f'benchopt_run_{timestamp}.parquet'
     else:
-        save_file = output_dir / f"{output_name}.csv"
+        save_file = output_dir / f"{output_name}.parquet"
         save_file = uniquify_results(save_file)
-    df.to_csv(save_file)
+    df.to_parquet(save_file)
     output.savefile_status(save_file=save_file)
 
     if plot_result: