+++ benchopt/tests/test_cli.py
@@ -1,4 +1,6 @@
 import re
+import time
+import tarfile
 import tempfile
 from pathlib import Path
 
@@ -24,8 +26,10 @@ from benchopt.tests import REQUIREMENT_BENCHMARK_PATH
 from benchopt.cli.main import run
 from benchopt.cli.main import install
 from benchopt.cli.helpers import clean
-from benchopt.cli.process_results import plot
+from benchopt.cli.helpers import archive
 from benchopt.cli.helpers import check_install
+from benchopt.cli.process_results import plot
+from benchopt.cli.process_results import generate_results
 
 
 BENCHMARK_COMPLETION_CASES = [
@@ -296,7 +300,9 @@ class TestRunCmd:
             run(command, 'benchopt', standalone_mode=False)
             run(command, 'benchopt', standalone_mode=False)
 
-        result_files = re.findall(r'Saving result in: (.*\.csv)', out.output)
+        result_files = re.findall(
+            r'Saving result in: (.*\.parquet)', out.output
+        )
         names = [Path(result_file).stem for result_file in result_files]
         assert names[0] == 'unique_name' and names[1] == 'unique_name_1'
 
@@ -413,10 +419,13 @@ class TestPlotCmd:
                  '-s', SELECT_ONE_PGD, '-n', '2', '-r', '1', '-o',
                  SELECT_ONE_OBJECTIVE, '--no-plot'], 'benchopt',
                 standalone_mode=False)
-        result_files = re.findall(r'Saving result in: (.*\.csv)', out.output)
+        result_files = re.findall(
+            r'Saving result in: (.*\.parquet)', out.output
+        )
         assert len(result_files) == 1, out.output
         result_file = result_files[0]
         cls.result_file = result_file
+        cls.result_file = str(Path(result_file).relative_to(Path().resolve()))
 
     @classmethod
     def teardown_class(cls):
@@ -426,15 +435,20 @@ class TestPlotCmd:
     def test_plot_invalid_file(self):
 
         with pytest.raises(FileNotFoundError, match=r"invalid_file"):
-            plot([str(DUMMY_BENCHMARK_PATH), '-f', 'invalid_file'],
-                 'benchopt', standalone_mode=False)
+            plot([str(DUMMY_BENCHMARK_PATH), '-f', 'invalid_file', '--no-html',
+                  '--no-display'], 'benchopt', standalone_mode=False)
 
     def test_plot_invalid_kind(self):
 
         with pytest.raises(ValueError, match=r"invalid_kind"):
+            plot([str(DUMMY_BENCHMARK_PATH), '-k', 'invalid_kind', '--no-html',
+                  '--no-display'], 'benchopt', standalone_mode=False)
 
-            plot([str(DUMMY_BENCHMARK_PATH), '-k', 'invalid_kind'],
-                 'benchopt', standalone_mode=False)
+    def test_plot_html_ignore_kind(self):
+
+        with pytest.warns(UserWarning, match=r"Cannot specify '--kind'"):
+            plot([str(DUMMY_BENCHMARK_PATH), '-k', 'invalid_kind', '--html',
+                  '--no-display'], 'benchopt', standalone_mode=False)
 
     @pytest.mark.parametrize('kind', PLOT_KINDS)
     def test_valid_call(self, kind):
@@ -443,12 +457,31 @@ class TestPlotCmd:
             plot([str(DUMMY_BENCHMARK_PATH), '-f', self.result_file,
                   '-k', kind, '--no-display', '--no-html'],
                  'benchopt', standalone_mode=False)
-        saved_files = re.findall(r'Save .* as: (.*\.pdf)', out.output)
-        assert len(saved_files) == 1
-        saved_file = saved_files[0]
-        assert kind in saved_file
 
-        Path(saved_file).unlink()
+        saved_files = re.findall(r'Save .* as: (.*\.pdf)', out.output)
+        try:
+            assert len(saved_files) == 1
+            assert kind in saved_files[0]
+        finally:
+            # Make sure to clean up all files even when the test fails
+            for f in saved_files:
+                Path(f).unlink()
+
+    def test_valid_call_html(self):
+
+        with SuppressStd() as out:
+            plot([str(DUMMY_BENCHMARK_PATH), '-f', self.result_file,
+                  '--no-display', '--html'], 'benchopt', standalone_mode=False)
+
+        saved_files = re.findall(
+            r'Writing.* results to (.*\.html)', out.output
+        )
+        try:
+            assert len(saved_files) == 2
+        finally:
+            # Make sure to clean up all files even when the test fails
+            for f in saved_files:
+                Path(f).unlink()
 
     def test_shell_complete(self):
         # Completion for benchmark name
@@ -456,10 +489,168 @@ class TestPlotCmd:
 
         # Completion for solvers
         _test_shell_completion(
-            run, [str(DUMMY_BENCHMARK_PATH), '-s'], SOLVER_COMPLETION_CASES
+            plot, [str(DUMMY_BENCHMARK_PATH), '-f'], [
+                ('', [self.result_file]),
+                (self.result_file[:-4], [self.result_file]),
+                ('_invalid_file', []),
+            ]
         )
 
-        # Completion for datasets
-        _test_shell_completion(
-            run, [str(DUMMY_BENCHMARK_PATH), '-d'], DATASET_COMPLETION_CASES
+
+class TestGenerateResultCmd:
+
+    @classmethod
+    def setup_class(cls):
+        "Make sure at least one result file is available"
+        with SuppressStd() as out:
+            clean([str(DUMMY_BENCHMARK_PATH)],
+                  'benchopt', standalone_mode=False)
+            clean([str(REQUIREMENT_BENCHMARK_PATH)],
+                  'benchopt', standalone_mode=False)
+            run([str(DUMMY_BENCHMARK_PATH), '-l', '-d', SELECT_ONE_SIMULATED,
+                 '-s', SELECT_ONE_PGD, '-n', '2', '-r', '1', '-o',
+                 SELECT_ONE_OBJECTIVE, '--no-plot'], 'benchopt',
+                standalone_mode=False)
+            time.sleep(1)  # Make sure there is 2 separate files
+            run([str(DUMMY_BENCHMARK_PATH), '-l', '-d', SELECT_ONE_SIMULATED,
+                 '-s', SELECT_ONE_PGD, '-n', '2', '-r', '1', '-o',
+                 SELECT_ONE_OBJECTIVE, '--no-plot'], 'benchopt',
+                standalone_mode=False)
+        result_files = re.findall(
+            r'Saving result in: (.*\.parquet)', out.output
         )
+        assert len(result_files) == 2, out.output
+        cls.result_files = result_files
+
+    @classmethod
+    def teardown_class(cls):
+        "Make sure at least one result file is available"
+        for f in cls.result_files:
+            Path(f).unlink()
+
+    def test_call(self):
+
+        with SuppressStd() as out:
+            generate_results([
+                '--root', str(DUMMY_BENCHMARK_PATH.parent), '--no-display'
+            ], 'benchopt', standalone_mode=False)
+        html_results = re.findall(r'Writing results to (.*\.html)', out.output)
+        html_benchmark = re.findall(
+            rf'Writing {DUMMY_BENCHMARK.name} results to (.*\.html)',
+            out.output
+        )
+        html_index = re.findall(r'Writing index to (.*\.html)', out.output)
+        try:
+            assert len(html_index) == 1, out.output
+            assert len(html_benchmark) == 1, out.output
+            assert len(html_results) == len(self.result_files), out.output
+            print(out.output)
+            for f in self.result_files:
+                basename = Path(f).stem
+                assert any(basename in res for res in html_results)
+        finally:
+            # Make sure to clean up all files even when the test fails
+            for f in html_results + html_benchmark + html_index:
+                Path(f).unlink()
+
+
+class TestArchiveCmd:
+
+    @classmethod
+    def setup_class(cls):
+        "Make sure at least one result file is available"
+        with SuppressStd() as out:
+            run([str(DUMMY_BENCHMARK_PATH), '-l', '-d', SELECT_ONE_SIMULATED,
+                 '-s', SELECT_ONE_PGD, '-n', '2', '-r', '1', '-o',
+                 SELECT_ONE_OBJECTIVE, '--no-plot'], 'benchopt',
+                standalone_mode=False)
+        result_file = re.findall(
+            r'Saving result in: (.*\.parquet)', out.output
+        )
+        assert len(result_file) == 1, out.output
+        cls.result_file = result_file[0]
+
+    @classmethod
+    def teardown_class(cls):
+        "Clean up the result file."
+        Path(cls.result_file).unlink()
+
+    @pytest.mark.parametrize('invalid_benchmark, match', [
+        ('invalid_benchmark', "Path 'invalid_benchmark' does not exist."),
+        ('.', "The folder '.' does not contain `objective.py`"),
+        ("", rf"The folder '{CURRENT_DIR}' does not contain `objective.py`")],
+        ids=['invalid_path', 'no_objective', "no_objective in default"])
+    def test_invalid_benchmark(self, invalid_benchmark, match):
+        with pytest.raises(click.BadParameter, match=match):
+            if len(invalid_benchmark) > 0:
+                run([invalid_benchmark], 'benchopt', standalone_mode=False)
+            else:
+                run([], 'benchopt', standalone_mode=False)
+
+    def test_call(self):
+
+        with SuppressStd() as out:
+            archive([str(DUMMY_BENCHMARK_PATH)], 'benchopt',
+                    standalone_mode=False)
+        saved_files = re.findall(r'Results are in (.*\.tar.gz)', out.output)
+        try:
+            assert len(saved_files) == 1
+            saved_file = saved_files[0]
+
+            counts = {k: 0 for k in [
+                "__pycache__", "outputs", "objective.py", "datasets",
+                "solvers", "README"
+            ]}
+
+            with tarfile.open(saved_file, "r:gz") as tar:
+                for elem in tar.getmembers():
+                    for k in counts:
+                        counts[k] += k in elem.name
+                    assert elem.uname == "benchopt"
+
+            assert counts["README"] == 1, counts
+            assert counts["objective.py"] == 1, counts
+            assert counts["datasets"] >= 1, counts
+            assert counts["solvers"] >= 1, counts
+            assert counts["outputs"] == 0, counts
+            assert counts["__pycache__"] == 0, counts
+        finally:
+            # Make sure to clean up all files even when the test fails
+            for f in saved_files:
+                Path(f).unlink()
+
+    def test_call_with_outputs(self):
+
+        with SuppressStd() as out:
+            archive([str(DUMMY_BENCHMARK_PATH), "--with-outputs"], 'benchopt',
+                    standalone_mode=False)
+        saved_files = re.findall(r'Results are in (.*\.tar.gz)', out.output)
+        try:
+            assert len(saved_files) == 1
+            saved_file = saved_files[0]
+
+            counts = {k: 0 for k in [
+                "__pycache__", "outputs", "objective.py", "datasets",
+                "solvers", "README"
+            ]}
+
+            with tarfile.open(saved_file, "r:gz") as tar:
+                for elem in tar.getmembers():
+                    for k in counts:
+                        counts[k] += k in elem.name
+                    assert elem.uname == "benchopt"
+
+            assert counts["README"] == 1, counts
+            assert counts["objective.py"] == 1, counts
+            assert counts["datasets"] >= 1, counts
+            assert counts["solvers"] >= 1, counts
+            assert counts["outputs"] >= 1, counts
+            assert counts["__pycache__"] == 0, counts
+        finally:
+            # Make sure to clean up all files even when the test fails
+            for f in saved_files:
+                Path(f).unlink()
+
+    def test_shell_complete(self):
+        # Completion for benchmark name
+        _test_shell_completion(archive, [], BENCHMARK_COMPLETION_CASES)