+++ benchopt/utils/files.py
@@ -1,5 +1,5 @@
-from pathlib import Path
 import warnings
+from pathlib import Path
 
 
 def rm_folder(folder):
@@ -14,7 +14,7 @@ def rm_folder(folder):
 
 
 def uniquify_results(file_path):
-    "Add a number to filename if it already exists"
+    "Add a number to filename if it already exists."
     if file_path.exists():
         parent = file_path.parent
         stem = file_path.stem
@@ -25,7 +25,7 @@ def uniquify_results(file_path):
         alternative = parent / f"{stem}_{i}{suffix}"
         warnings.warn(
             f"{file_path} already exists. Saving results to {alternative}"
-            )
+        )
         return alternative
     else:
         return file_path