+++ benchopt/utils/sys_info.py
@@ -1,11 +1,15 @@
 import io
 import os
 import re
+import warnings
 import platform
 import contextlib
 import subprocess
 from shutil import which
 from pathlib import Path
+
+from ..config import DEBUG
+from .shell_cmd import _run_shell
 
 
 def _get_processor_name():
@@ -33,7 +37,6 @@ def get_cuda_version():
     try:
         out = subprocess.check_output(command).strip().decode("utf-8")
     except subprocess.CalledProcessError:
-        import warnings
         warnings.warn(
             "`nvidia-smi` has failed. Please check NVIDIA driver install."
         )
@@ -43,10 +46,9 @@ def get_cuda_version():
         name = re.search('<product_name>(.*)</product_name>', out).group(1)
         return f"{name}: cuda_{version}"
     except AttributeError:
-        import warnings
         warnings.warn(
             "Could not parse cuda version or device name from `nvidia-smi`."
-         )
+        )
         return None
 
 
@@ -74,6 +76,15 @@ def _get_numpy_libs():
                 libs += [f"{key}={lib}"]
     libs = ", ".join(libs)
     return libs
+
+
+def _get_git_tag():
+    err, tag = _run_shell("git describe --tags --abbrev=0", return_output=True)
+    if err != 0:
+        if DEBUG:
+            print(err, tag)
+        tag = None
+    return tag
 
 
 def get_sys_info():
@@ -107,5 +118,8 @@ def get_sys_info():
     info["version-cuda"] = get_cuda_version()
     info["version-numpy"] = (np.__version__, _get_numpy_libs())
     info["version-scipy"] = scipy.__version__
+
+    # Info on benchmark version
+    info["benchmark-git-tag"] = _get_git_tag()
 
     return info